requirements.txt
seaborn==0.11.2
pandas==1.3.4
matplotlib==3.4.3
Eel==0.14.0
speedtest-cli==2.1.3
openpyxl==3.0.9